collector/requirements.txt
pydantic>=2.0
pyyaml>=6.0
httpx[socks]>=0.27
feedparser>=6.0
openai>=1.0
python-dateutil>=2.8
